requirements/requirements.txt
matplotlib==3.10.5
numpy==2.3.2
Pillow==11.3.0
scikit_learn==1.7.1
seaborn==0.13.2
torch==2.7.1+cu118
torchvision==0.22.1+cu118
